create mixin for BoardColumn and ColumnTask
